requirements.txt
requests>=2.31.0
flask>=3.0.0
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7
